+++ R/install_beast2.R
@@ -49,7 +49,8 @@ install_beast2 <- function(
   local_path <- file.path(folder_name, archive_filename)
   utils::download.file(
     url = url,
-    destfile = local_path
+    destfile = local_path,
+    quiet = !verbose
   )
   beautier::check_file_exists(local_path, "local_path")